Create Esmioter Delta
Esmioter Delta is 2 times larger than earth and has 3 moons: Selia, Meztli, and Jana. Besides the peaceful humannoid natives that live ther now, Terra-forming may be the only way to ever bring human life to this planet. The current conditions are far too extreme for most lifeforms, but with the right technologies, life could be introduced. The planet also has plenty of resources to allow for these construction to be done on the planet.